# gcs: Optimize Hash.