**This is rebased on top of #1851**.

This adds a benchmark and optimizes the `Hash` method of gcs filters by making use of the new zero-alloc hashing func available in `crypto/blake256`.

The following is a before and after comparison:

```
benchmark       old ns/op   new ns/op    delta
-------------------------------------------------
BenchmarkHash   1786        1315         -26.37%

benchmark       old allocs  new allocs   delta
-------------------------------------------------
BenchmarkHash   2           0            -100.00%

benchmark       old bytes   new bytes    delta
-------------------------------------------------
BenchmarkHash   176         0            -100.00%
```